Refresh Interactor Test
